feat(util): 🎸 create line-by-line patch module
